test: merge customer-merge.cy.ts into customer-duplicates.cy.ts
Both specs cover one connected flow (the duplicates list navigates
straight into the merge picker), so keep them in one file with two
describe blocks - same pattern general.cy.ts already uses - and share
one createDuplicatePair() helper instead of two near-identical ones.
