+++ frontend/src/main/webapp/cypress/e2e/customer-duplicates.cy.ts
@@ -9,7 +9,7 @@ describe('Customer Duplicates', () => {
   });
 
   it('lists a detected duplicate pair with both customers\' details', () => {
-    createDuplicateCustomerPair().then(({first, second}) => {
+    createDuplicatePair().then(({first, second}) => {
       const customer1 = first.body.data;
       const customer2 = second.body.data;
 
@@ -32,7 +32,7 @@ describe('Customer Duplicates', () => {
   });
 
   it('navigates to the customer detail page', () => {
-    createDuplicateCustomerPair().then(({first}) => {
+    createDuplicatePair().then(({first}) => {
       const customer1 = first.body.data;
 
       cy.visit('/#/kunden/duplikate');
@@ -43,7 +43,7 @@ describe('Customer Duplicates', () => {
   });
 
   it('deletes a single customer from the duplicate pair', () => {
-    createDuplicateCustomerPair().then(({second}) => {
+    createDuplicatePair().then(({second}) => {
       const customer2 = second.body.data;
 
       cy.visit('/#/kunden/duplikate');
@@ -55,7 +55,7 @@ describe('Customer Duplicates', () => {
   });
 
   it('opens the merge picker for the duplicate pair', () => {
-    createDuplicateCustomerPair().then(({first, second}) => {
+    createDuplicatePair().then(({first, second}) => {
       const customer1 = first.body.data;
       const customer2 = second.body.data;
 
@@ -70,7 +70,7 @@ describe('Customer Duplicates', () => {
   it('stacks a duplicate pair into a single column on phone and deletion still works', () => {
     cy.viewport(PHONE_VIEWPORT);
 
-    createDuplicateCustomerPair().then(({first, second}) => {
+    createDuplicatePair().then(({first, second}) => {
       const customer1 = first.body.data;
       const customer2 = second.body.data;
 
@@ -99,7 +99,7 @@ describe('Customer Duplicates', () => {
   it('renders a duplicate pair side-by-side at tablet breakpoint and merge still works', () => {
     cy.viewport(TABLET_VIEWPORT);
 
-    createDuplicateCustomerPair().then(({first, second}) => {
+    createDuplicatePair().then(({first, second}) => {
       const customer1 = first.body.data;
       const customer2 = second.body.data;
 
@@ -121,7 +121,144 @@ describe('Customer Duplicates', () => {
 
 });
 
-function createDuplicateCustomerPair() {
+describe('Customer Merge', () => {
+
+  beforeEach(() => {
+    cy.loginDefault();
+  });
+
+  it('picks a conflicting field from the source and applies it to the target', () => {
+    createDuplicatePair({telephoneNumber: '111111'}, {telephoneNumber: '222222'}).then(({first, second}) => {
+      const target = first.body.data;
+      const source = second.body.data;
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+
+      cy.url().should('include', `/kunden/zusammenfuehren/${target.id}`);
+      cy.byTestId('merge-field-TELEPHONE_NUMBER-source-' + source.id).click();
+      cy.byTestId('merge-confirm-button').click();
+
+      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
+      cy.url().should('include', '/kunden/detail/' + target.id);
+      cy.byTestId('telephoneNumberText').should('contain.text', '222222');
+    });
+  });
+
+  it('re-parents an additional person that only exists on the source', () => {
+    createDuplicatePair({}, {
+      additionalPersons: [{
+        key: 0,
+        id: 0,
+        firstname: 'Peter',
+        lastname: 'Novak',
+        birthDate: dayjs().subtract(40, 'year').toDate(),
+        gender: Gender.MALE,
+        country: {id: 165, code: 'AT', name: 'Österreich'},
+        excludeFromHousehold: false,
+        receivesFamilyAllowance: false
+      }]
+    }).then(({first}) => {
+      const target = first.body.data;
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+      cy.byTestId('merge-confirm-button').click();
+
+      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
+      cy.url().should('include', '/kunden/detail/' + target.id);
+      cy.byTestId('addperson-0-lastnameText').should('contain.text', 'Novak');
+    });
+  });
+
+  it('deduplicates an additional person that exists identically on both target and source', () => {
+    const sharedPerson = (): any => ({
+      key: 0,
+      id: 0,
+      firstname: 'Anna',
+      lastname: 'Schmidt',
+      birthDate: dayjs().subtract(35, 'year').toDate(),
+      gender: Gender.FEMALE,
+      country: {id: 165, code: 'AT', name: 'Österreich'},
+      excludeFromHousehold: false,
+      receivesFamilyAllowance: false
+    });
+
+    createDuplicatePair(
+      {additionalPersons: [sharedPerson()]},
+      {additionalPersons: [sharedPerson()]}
+    ).then(({first}) => {
+      const target = first.body.data;
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+      cy.byTestId('merge-confirm-button').click();
+
+      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
+      cy.url().should('include', '/kunden/detail/' + target.id);
+      cy.byTestId('addperson-0-lastnameText').should('contain.text', 'Schmidt');
+      cy.byTestId('addperson-1-lastnameText').should('not.exist');
+    });
+  });
+
+  it('re-parents a note from the source onto the target', () => {
+    createDuplicatePair().then(({first, second}) => {
+      const target = first.body.data;
+      const source = second.body.data;
+
+      cy.createCustomerNote(source.id!, 'note from the merged-away source');
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+      cy.byTestId('merge-confirm-button').click();
+
+      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
+      cy.url().should('include', '/kunden/detail/' + target.id);
+      cy.byTestId('note-text').should('contain.text', 'note from the merged-away source');
+    });
+  });
+
+  it('warns about a same-distribution ticket collision and still lets the merge succeed', () => {
+    cy.createDistribution();
+
+    createDuplicatePair().then(({first, second}) => {
+      const target = first.body.data;
+      const source = second.body.data;
+
+      cy.addCustomerToDistribution({customerId: target.id!, ticketNumber: 11});
+      cy.addCustomerToDistribution({customerId: source.id!, ticketNumber: 22});
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+
+      cy.contains('22').should('be.visible');
+      cy.contains('wird verworfen').should('be.visible');
+
+      cy.byTestId('merge-confirm-button').click();
+      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
+
+      cy.closeDistribution();
+    });
+  });
+
+  it('cancel returns to the duplicates list without merging', () => {
+    createDuplicatePair().then(({first, second}) => {
+      const target = first.body.data;
+      const source = second.body.data;
+
+      cy.visit('/#/kunden/duplikate');
+      cy.byTestId('duplicate-merge-button-' + target.id).click();
+      cy.byTestId('merge-cancel-button').click();
+
+      cy.url().should('include', '/kunden/duplikate');
+      cy.byTestId('duplicate-customer-' + target.id).should('exist');
+      cy.byTestId('duplicate-customer-' + source.id).should('exist');
+    });
+  });
+
+});
+
+function createDuplicatePair(firstOverrides: Partial<CustomerData> = {}, secondOverrides: Partial<CustomerData> = {}) {
   return cy.getAnyRandomNumber().then(randomNumber => {
     const buildCustomer = (): CustomerData => ({
       firstname: 'Firstname' + randomNumber,
@@ -144,8 +281,8 @@ function createDuplicateCustomerPair() {
       validUntil: dayjs().add(1, 'year').toDate()
     });
 
-    return cy.createCustomer(buildCustomer()).then(first =>
-      cy.createCustomer(buildCustomer()).then(second => ({first, second}))
+    return cy.createCustomer({...buildCustomer(), ...firstOverrides}).then(first =>
+      cy.createCustomer({...buildCustomer(), ...secondOverrides}).then(second => ({first, second}))
     );
   });
 }
