--- frontend/src/main/webapp/cypress/e2e/customer-merge.cy.ts
@@ -1,168 +1,0 @@
-import dayjs from 'dayjs';
-import {CustomerData, Gender} from '../support/commands';
-
-describe('Customer Merge', () => {
-
-  beforeEach(() => {
-    cy.loginDefault();
-  });
-
-  it('picks a conflicting field from the source and applies it to the target', () => {
-    createDuplicatePair({telephoneNumber: '111111'}, {telephoneNumber: '222222'}).then(({first, second}) => {
-      const target = first.body.data;
-      const source = second.body.data;
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-
-      cy.url().should('include', `/kunden/zusammenfuehren/${target.id}`);
-      cy.byTestId('merge-field-TELEPHONE_NUMBER-source-' + source.id).click();
-      cy.byTestId('merge-confirm-button').click();
-
-      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
-      cy.url().should('include', '/kunden/detail/' + target.id);
-      cy.byTestId('telephoneNumberText').should('contain.text', '222222');
-    });
-  });
-
-  it('re-parents an additional person that only exists on the source', () => {
-    createDuplicatePair({}, {
-      additionalPersons: [{
-        key: 0,
-        id: 0,
-        firstname: 'Peter',
-        lastname: 'Novak',
-        birthDate: dayjs().subtract(40, 'year').toDate(),
-        gender: Gender.MALE,
-        country: {id: 165, code: 'AT', name: 'Österreich'},
-        excludeFromHousehold: false,
-        receivesFamilyAllowance: false
-      }]
-    }).then(({first}) => {
-      const target = first.body.data;
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-      cy.byTestId('merge-confirm-button').click();
-
-      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
-      cy.url().should('include', '/kunden/detail/' + target.id);
-      cy.byTestId('addperson-0-lastnameText').should('contain.text', 'Novak');
-    });
-  });
-
-  it('deduplicates an additional person that exists identically on both target and source', () => {
-    const sharedPerson = (): any => ({
-      key: 0,
-      id: 0,
-      firstname: 'Anna',
-      lastname: 'Schmidt',
-      birthDate: dayjs().subtract(35, 'year').toDate(),
-      gender: Gender.FEMALE,
-      country: {id: 165, code: 'AT', name: 'Österreich'},
-      excludeFromHousehold: false,
-      receivesFamilyAllowance: false
-    });
-
-    createDuplicatePair(
-      {additionalPersons: [sharedPerson()]},
-      {additionalPersons: [sharedPerson()]}
-    ).then(({first}) => {
-      const target = first.body.data;
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-      cy.byTestId('merge-confirm-button').click();
-
-      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
-      cy.url().should('include', '/kunden/detail/' + target.id);
-      cy.byTestId('addperson-0-lastnameText').should('contain.text', 'Schmidt');
-      cy.byTestId('addperson-1-lastnameText').should('not.exist');
-    });
-  });
-
-  it('re-parents a note from the source onto the target', () => {
-    createDuplicatePair({}, {}).then(({first, second}) => {
-      const target = first.body.data;
-      const source = second.body.data;
-
-      cy.createCustomerNote(source.id!, 'note from the merged-away source');
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-      cy.byTestId('merge-confirm-button').click();
-
-      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
-      cy.url().should('include', '/kunden/detail/' + target.id);
-      cy.byTestId('note-text').should('contain.text', 'note from the merged-away source');
-    });
-  });
-
-  it('warns about a same-distribution ticket collision and still lets the merge succeed', () => {
-    cy.createDistribution();
-
-    createDuplicatePair({}, {}).then(({first, second}) => {
-      const target = first.body.data;
-      const source = second.body.data;
-
-      cy.addCustomerToDistribution({customerId: target.id!, ticketNumber: 11});
-      cy.addCustomerToDistribution({customerId: source.id!, ticketNumber: 22});
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-
-      cy.contains('22').should('be.visible');
-      cy.contains('wird verworfen').should('be.visible');
-
-      cy.byTestId('merge-confirm-button').click();
-      cy.get('.toast-message').should('be.visible').and('contain.text', 'zusammengeführt');
-
-      cy.closeDistribution();
-    });
-  });
-
-  it('cancel returns to the duplicates list without merging', () => {
-    createDuplicatePair({}, {}).then(({first, second}) => {
-      const target = first.body.data;
-      const source = second.body.data;
-
-      cy.visit('/#/kunden/duplikate');
-      cy.byTestId('duplicate-merge-button-' + target.id).click();
-      cy.byTestId('merge-cancel-button').click();
-
-      cy.url().should('include', '/kunden/duplikate');
-      cy.byTestId('duplicate-customer-' + target.id).should('exist');
-      cy.byTestId('duplicate-customer-' + source.id).should('exist');
-    });
-  });
-
-});
-
-function createDuplicatePair(firstOverrides: Partial<CustomerData>, secondOverrides: Partial<CustomerData>) {
-  return cy.getAnyRandomNumber().then(randomNumber => {
-    const buildCustomer = (): CustomerData => ({
-      firstname: 'Firstname' + randomNumber,
-      lastname: 'Lastname' + randomNumber,
-      birthDate: dayjs().subtract(30, 'year').toDate(),
-      gender: Gender.MALE,
-      country: {
-        id: 165,
-        code: 'AT',
-        name: 'Österreich'
-      },
-      income: 1000,
-      incomeDue: dayjs().add(30, 'days').toDate(),
-      address: {
-        street: 'Duplicatestreet' + randomNumber,
-        houseNumber: '1',
-        city: 'city-' + randomNumber,
-        postalCode: 1234
-      },
-      validUntil: dayjs().add(1, 'year').toDate()
-    });
-
-    return cy.createCustomer({...buildCustomer(), ...firstOverrides}).then(first =>
-      cy.createCustomer({...buildCustomer(), ...secondOverrides}).then(second => ({first, second}))
-    );
-  });
-}
